+++ cr_channable_connector/doc/changelog.rst
@@ -1,4 +1,10 @@
 17.0.0.0(Date: 16th June,2026)
 -------------------------------
-
 - migrated the module.
+
+17.0.0.1(Date: 10th Aug,2026)
+-------------------------------
+- Update existing Odoo orders seamlessly during bulk sync processes.
+- Automatic Odoo order cancellation and credit note creation when a bulk sync detects a cancelled status in Channable.
+- Added 'Schedule Sync' smart button to the Channable Marketplace dashboard for easy sync interval configuration.
+- Removed 'Sync State' smart button because similar functionality like 'Sync Order'.